+++ src/CleanRegex/Match/IntStream.php
@@ -43,11 +43,6 @@ class IntStream implements \Countable, \IteratorAggregate
         return $this->terminal->all();
     }
 
-    public function only(int $limit): array
-    {
-        return $this->terminal->only($limit);
-    }
-
     public function forEach(callable $consumer): void
     {
         $this->terminal->forEach($consumer);